services/verification/main.py
"""
Verification Service - Ensures zero-hallucination responses by validating
LLM outputs against authoritative data sources.
"""

import os
from contextlib import asynccontextmanager
from typing import AsyncGenerator

import structlog
from dotenv import load_dotenv
from fastapi import FastAPI, HTTPException
from pydantic import BaseModel

load_dotenv()

logger = structlog.get_logger()


class VerificationRequest(BaseModel):
    """Request to verify an LLM response."""

    session_id: str
    response_text: str
    claimed_sources: list[str] = []
    context: dict | None = None


class VerificationResult(BaseModel):
    """Result of verification check."""

    verified: bool
    confidence: float
    citations: list[dict]
    warnings: list[str] = []
    modified_response: str | None = None


class SourceCheck(BaseModel):
    """Individual source verification."""

    source: str
    found: bool
    snippet: str | None = None
    url: str | None = None


@asynccontextmanager
async def lifespan(app: FastAPI) -> AsyncGenerator[None, None]:
    """Application lifespan manager."""
    logger.info("verification_service_starting")
    yield
    logger.info("verification_service_stopping")


app = FastAPI(
    title="Jarvis Verification Service",
    description="Validates LLM responses against authoritative sources",
    version="0.1.0",
    lifespan=lifespan,
)


@app.get("/health")
async def health_check() -> dict:
    """Health check endpoint."""
    return {"status": "ok", "service": "verification"}


@app.post("/verify", response_model=VerificationResult)
async def verify_response(request: VerificationRequest) -> VerificationResult:
    """
    Verify an LLM response for accuracy.

    This endpoint checks:
    1. All claimed facts have supporting sources
    2. Sources are authoritative and recent
    3. Response doesn't contain hallucinated information
    """
    logger.info(
        "verification_request",
        session_id=request.session_id,
        response_length=len(request.response_text),
    )

    # TODO: Implement actual verification logic
    # - Cross-reference with vector store
    # - Check API data cache
    # - Validate GitHub source citations

    # Placeholder implementation
    citations = []
    warnings = []

    for source in request.claimed_sources:
        # TODO: Actually verify each source
        citations.append(
            {
                "source": source,
                "verified": True,
                "snippet": "Source verification pending implementation",
            }
        )

    # For now, return unverified with a warning
    return VerificationResult(
        verified=False,
        confidence=0.0,
        citations=citations,
        warnings=["Verification service not fully implemented"],
        modified_response=None,
    )


@app.post("/check-source", response_model=SourceCheck)
async def check_source(source: str) -> SourceCheck:
    """Check if a specific source exists and is valid."""
    logger.info("source_check", source=source)

    # TODO: Implement source checking
    # - GitHub API lookup
    # - API cache lookup
    # - Vector store search

    return SourceCheck(
        source=source,
        found=False,
        snippet=None,
        url=None,
    )


if __name__ == "__main__":
    import uvicorn

    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=int(os.getenv("PORT", "8003")),
        reload=os.getenv("ENV", "development") == "development",
    )
