services/verification/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
httpx==0.26.0
openai==1.9.0
numpy==1.26.3
python-dotenv==1.0.0
redis==5.0.1
structlog==24.1.0
